Add optional fallback for `auto` and `engine`
This is a sketch commit that makes it possible to use, eg, `auto:lts` to
mean "same as `auto`, or `lts` if that didn't work". With this you can
get the benefit of `auto` when applicable, or `lts` otherwise.

An alternative approach would be to use an envvar, as in

    N_FALLBACK_VERSION=lts n auto

but that would be less convenient.

---

An example of a cute hack is to create an executable file that contains

    #!/bin/bash
    exec n -q -d exec auto:lts "$(basename "$0")" "$@"

then symlink it as `/usr/local/bin/node` etc, and now you have a magical
`node` command that follows whatever version is specified in your
current directory. (This fails in case of running with `engine`, since
then requiring node itself means an infinite loop, but could be done
with some more tweaks.)