Add `$N_ARCH` as a default architecture
More convenient to use than --arch in case of a mirror with exotic
architecture, like the ones in `unofficial-builds`.

Do this by making it the default for `$ARCH`.

Also, skip the `uname -m` if it's not needed because of this, or because
of `--arch`, or the version-specific default. (This is not needed for
the functionality.)